requirements.txt
python-gnupg
pyyaml
asfpy>=0.37
requests
